Fixes wrong display of layout for non-prose content
The pages that didn't use the prose display (`TIL`, `/` and `About me`) had a wrong
display because Tailwind was not applying `max-w-5xl`. This commit adds it to
tailwind.config file to ensure it can be applied.

Also, taking advantage to reformat a bit the display of subtitle in articles, so it
looks less overwhelming.